requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
gunicorn>=20.1,<21.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=3.13,<4.0
django-environ>=0.4.5
django-import-export
django-filter
django-allauth
dj-rest-auth[with_social]==5.0.0
djangorestframework-simplejwt==5.3.1
django-storages[s3]
shortuuid
pillow